onvif: Use double for absoluteMove and relativeMove arguments
The wrapper methods absoluteMove() and relativeMove() declared their
pan/tilt/zoom arguments as int, even though ONVIF velocities are
normalized doubles in the range [-1.0, 1.0] (and zoom in [0.0, 1.0]).
Every absolute or relative move was therefore truncating to 0 or 1
before being serialized into the SOAP envelope.

The underlying genericMove() already takes doubles, so this is a pure
signature fix with no behavioral risk for any code path that already
worked (whole-range moves still go to 1).

+++ src/ptz-onvif.cpp
@@ -137,12 +137,12 @@ void PTZOnvif::continuousMove(double x, double y, double z)
 	genericMove("ContinuousMove", "Velocity", x, y, z);
 }
 
-void PTZOnvif::absoluteMove(int x, int y, int z)
+void PTZOnvif::absoluteMove(double x, double y, double z)
 {
 	genericMove("AbsoluteMove", "Position", x, y, z);
 }
 
-void PTZOnvif::relativeMove(int x, int y, int z)
+void PTZOnvif::relativeMove(double x, double y, double z)
 {
 	genericMove("RelativeMove", "Translation", x, y, z);
 }
