+++ src/ptz-onvif.hpp
@@ -55,8 +55,8 @@ private:
 
 	void genericMove(QString movetype, QString property, double pan, double tilt, double zoom);
 	void continuousMove(double x, double y, double z);
-	void absoluteMove(int x, int y, int z);
-	void relativeMove(int x, int y, int z);
+	void absoluteMove(double x, double y, double z);
+	void relativeMove(double x, double y, double z);
 	void stop();
 	void goToHomePosition();
